Handlers, DB queries refactoring
